Fixed implicit cast

+++ src/Error.cpp
@@ -3,11 +3,11 @@
 #include "fmt/core.h"
 
 namespace ErrorReporter {
-    void Error(int line, const std::string& message) {
+    void Error(size_t line, const std::string& message) {
         Report(line, "", message);
     }
 
-    void Report(int line, const std::string& where, const std::string& message) {
+    void Report(size_t line, const std::string& where, const std::string& message) {
         // Needs something like
         // Error: Unexpected "," in argument list.
         //
